split single file
