--- src/bot.ts
@@ -1,3 +1,0 @@
-import * as discordBot from "./discord-bot";
-
-discordBot.start();
