+++ src/start-discord-bot.ts
@@ -1,0 +1,3 @@
+import * as discordBot from "./discord/discord-bot";
+
+discordBot.start();
